Move importHtml tests into importText test file.
There is no importHtml. They are really importText tests.
